GDA-5883 ScanDataPoint scan number is now an int

+++ uk.ac.gda.epics.test/src/gda/device/detector/addetector/ADDetectorTest.java
@@ -90,7 +90,7 @@ public class ADDetectorTest {
 		scanInfo = mock(ScanInformation.class);
 		ICurrentScanInformationHolder currentScanHolder = mock(ICurrentScanInformationHolder.class);
 		when(currentScanHolder.getCurrentScanInformation()).thenReturn(scanInfo);
-		when(scanInfo.getScanNumber()).thenReturn((long) 12345);
+		when(scanInfo.getScanNumber()).thenReturn(12345);
 		InterfaceProvider.setCurrentScanInformationHolderForTesting(currentScanHolder);
 	}
